Separate logic for specific native components
Break up 'createStrictDOMComponent' into separate functions for
different types of React Native component. This can help to reduce the
amount of unused logic and hooks needed by each type of native
component.
